+++ js/utilits.js
@@ -1,0 +1,17 @@
+function getRandomNumber1 (first, second, digits = 1) {
+  const lower = Math.min(Math.abs(first), Math.abs(second));
+  const upper = Math.max(Math.abs(first), Math.abs(second));
+  const result = Math.random() * (upper - lower) + lower;
+
+  return result.toFixed(digits);
+}
+
+// function getRandomNumber (first, second) {
+//   const lower = Math.ceil(Math.min(Math.abs(first), Math.abs(second)));
+//   const upper = Math.floor(Math.max(Math.abs(first), Math.abs(second)));
+//   const result = Math.random() * (upper - lower + 1) + lower;
+
+//   return Math.floor(result);
+// }
+
+export {getRandomNumber1};
